reformat estimator_checks into module in tests folder (#421)
