+++ tests/test_dataframe_checks.py
@@ -85,9 +85,7 @@ def test_check_x_y_returns_pandas_from_pandas(df_vartypes):
 
 
 def test_check_X_y_returns_pandas_from_pandas_with_non_typical_index():
-    df = pd.DataFrame(
-        {"0": [1, 2, 3, 4], "1": [5, 6, 7, 8]}, index=[22, 99, 101, 212]
-    )
+    df = pd.DataFrame({"0": [1, 2, 3, 4], "1": [5, 6, 7, 8]}, index=[22, 99, 101, 212])
     s = pd.Series([1, 2, 3, 4], index=[22, 99, 101, 212])
     x, y = check_X_y(df, s)
     assert_frame_equal(df, x)
@@ -95,9 +93,7 @@ def test_check_X_y_returns_pandas_from_pandas_with_non_typical_index():
 
 
 def test_check_X_y_raises_error_when_pandas_index_dont_match():
-    df = pd.DataFrame(
-        {"0": [1, 2, 3, 4], "1": [5, 6, 7, 8]}, index=[22, 99, 101, 212]
-    )
+    df = pd.DataFrame({"0": [1, 2, 3, 4], "1": [5, 6, 7, 8]}, index=[22, 99, 101, 212])
     s = pd.Series([1, 2, 3, 4], index=[22, 99, 101, 999])
     with pytest.raises(ValueError):
         check_X_y(df, s)
@@ -105,9 +101,7 @@ def test_check_X_y_raises_error_when_pandas_index_dont_match():
 
 def test_check_x_y_reassings_index_when_only_one_input_is_pandas():
     # case 1: X is dataframe, y is something else
-    df = pd.DataFrame(
-        {"0": [1, 2, 3, 4], "1": [5, 6, 7, 8]}, index=[22, 99, 101, 212]
-    )
+    df = pd.DataFrame({"0": [1, 2, 3, 4], "1": [5, 6, 7, 8]}, index=[22, 99, 101, 212])
     s = np.array([1, 2, 3, 4])
     s_exp = pd.Series([1, 2, 3, 4], index=[22, 99, 101, 212])
     x, y = check_X_y(df, s)
