+++ tests/test_datetime/test_check_estimator_datetime.py
@@ -1,7 +1,7 @@
 import pytest
 
 from feature_engine.datetime import DatetimeFeatures
-from feature_engine.estimator_checks import check_feature_engine_estimator
+from tests.estimator_checks.estimator_checks import check_feature_engine_estimator
 
 _estimators = [DatetimeFeatures()]
