+++ tests/test_encoding/test_check_estimator_encoders.py
@@ -12,7 +12,7 @@ from feature_engine.encoding import (
     RareLabelEncoder,
     WoEEncoder,
 )
-from feature_engine.estimator_checks import check_feature_engine_estimator
+from tests.estimator_checks.estimator_checks import check_feature_engine_estimator
 
 _estimators = [
     CountFrequencyEncoder(ignore_format=True),
