+++ tests/test_imputation/test_check_estimator_imputers.py
@@ -1,7 +1,7 @@
 import pytest
 from sklearn.utils.estimator_checks import check_estimator
 
-from feature_engine.estimator_checks import check_feature_engine_estimator
+from tests.estimator_checks.estimator_checks import check_feature_engine_estimator
 from feature_engine.imputation import (
     AddMissingIndicator,
     ArbitraryNumberImputer,
