+++ tests/test_selection/test_check_estimator_selectors.py
@@ -2,9 +2,11 @@ import pytest
 from sklearn.linear_model import LogisticRegression
 from sklearn.utils.estimator_checks import check_estimator
 
-from feature_engine.estimator_checks import (
-    check_confirm_variables,
+from tests.estimator_checks.estimator_checks import (
     check_feature_engine_estimator,
+)
+from tests.estimator_checks.init_params_triggered_functionality_checks import (
+    check_confirm_variables,
 )
 from feature_engine.selection import (
     DropConstantFeatures,
