+++ tests/test_time_series/test_forecasting/test_check_estimator_forecasting.py
@@ -4,7 +4,7 @@ import pytest
 from sklearn.base import clone
 from sklearn.utils.estimator_checks import check_estimator
 
-from feature_engine.estimator_checks import check_feature_engine_estimator
+from tests.estimator_checks.estimator_checks import check_feature_engine_estimator
 from feature_engine.timeseries.forecasting import (
     ExpandingWindowFeatures,
     LagFeatures,
@@ -14,7 +14,7 @@ from feature_engine.timeseries.forecasting import (
 _estimators = [
     LagFeatures(missing_values="ignore"),
     WindowFeatures(missing_values="ignore"),
-    ExpandingWindowFeatures(missing_values="ignore")
+    ExpandingWindowFeatures(missing_values="ignore"),
 ]
