+++ tests/test_transformation/test_check_estimator_transformers.py
@@ -1,7 +1,7 @@
 import pytest
 from sklearn.utils.estimator_checks import check_estimator
 
-from feature_engine.estimator_checks import check_feature_engine_estimator
+from tests.estimator_checks.estimator_checks import check_feature_engine_estimator
 from feature_engine.transformation import (
     BoxCoxTransformer,
     LogCpTransformer,
